# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bmstu Open IT Platform",
        "title": "Forecast system",
        "contact": {
            "name": "API Support",
            "url": "https://vk.com/b.larkin",
            "email": "borislarkin18@mail.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/forecast/to_pred/{forecast_id}": {
            "post": {
                "description": "If there`s no draft found, a new draft is to be created.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preds_Forecs"
                ],
                "summary": "Add forecast to current user`s draft prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "forecast_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.PredictionWithForecasts"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/pr_fc/edit/{prediction_id}/{forecast_id}": {
            "put": {
                "description": "An error is returned in cases of unauthorized actions being attempted",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preds_Forecs"
                ],
                "summary": "Edit forecast`s input data for specified prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "forecast_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "prediction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdatePred_ForecInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Preds_Forecs"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/pr_fc/remove/{prediction_id}/{forecast_id}": {
            "delete": {
                "description": "An error is returned in cases of unauthorized actions being attempted",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preds_Forecs"
                ],
                "summary": "Delete forecast from specified prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "forecast_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "prediction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/login": {
            "post": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Login the specified user",
                "parameters": [
                    {
                        "description": "user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.loginResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Logout the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register the specified user",
                "parameters": [
                    {
                        "description": "New user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.registerResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/update/{id}": {
            "put": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update the specified user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.updateResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ds.ForecastResponseWithFlags": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "forecast_id",
                "image",
                "input",
                "measure_type",
                "result",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.Forecasts": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "forecast_id",
                "measure_type",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "ds.PredictionWithForecasts": {
            "type": "object",
            "required": [
                "forecasts",
                "prediction"
            ],
            "properties": {
                "forecasts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.ForecastResponseWithFlags"
                    }
                },
                "prediction": {
                    "$ref": "#/definitions/ds.Predictions"
                }
            }
        },
        "ds.Predictions": {
            "type": "object",
            "properties": {
                "date_completed": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "prediction_amount": {
                    "type": "integer"
                },
                "prediction_id": {
                    "type": "integer"
                },
                "prediction_window": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "ds.Preds_Forecs": {
            "type": "object",
            "required": [
                "forecast_id",
                "prediction_id",
                "preds_forecs_id"
            ],
            "properties": {
                "forecast": {
                    "$ref": "#/definitions/ds.Forecasts"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "input": {
                    "type": "string"
                },
                "prediction": {
                    "$ref": "#/definitions/ds.Predictions"
                },
                "prediction_id": {
                    "type": "integer"
                },
                "preds_forecs_id": {
                    "type": "integer"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "ds.Role": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "Guest",
                "User",
                "Moderator"
            ]
        },
        "ds.UpdatePred_ForecInput": {
            "type": "object",
            "required": [
                "input"
            ],
            "properties": {
                "input": {
                    "type": "string"
                }
            }
        },
...
        "handler.loginReq": {
            "type": "object",
            "required": [
                "guest",
                "login",
                "password"
            ],
            "properties": {
                "guest": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.loginResp": {
            "type": "object",
            "required": [
                "access_token",
                "expires_in",
                "login",
                "role",
                "token_type"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "handler.registerReq": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.registerResp": {
            "type": "object",
            "required": [
                "ok"
            ],
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "handler.updateReq": {
            "type": "object",
            "required": [
                "login",
                "password",
                "role"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/ds.Role"
                }
            }
        },
        "handler.updateResp": {
            "type": "object",
            "required": [
                "login",
                "role",
                "uid"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/ds.Role"
                },
                "uid": {
                    "type": "string"
                }
            }
        }
    }
}